requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
ase==3.20.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
biopython==1.78
black==20.8b1
bleach==3.2.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
dask==2020.12.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
googledrivedownloader==0.4
GPUtil==1.4.0
grpcio==1.34.0
h5py==3.1.0
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kaleido==0.1.0
kiwisolver==1.3.1
llvmlite==0.34.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numba==0.51.2
oauthlib==3.1.0
packaging==20.4
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.13.0
plyfile==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pykeops==1.4.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyVTK==0.5.18
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
rdflib==5.0.0
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.18.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
torch==1.10.0
torch-cluster
torch-geometric
torch-scatter
torch-sparse
torch-spline-conv
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
